WebSocket: queue close event as a task; stop wait_for_promise CPU spin
- didClose()/didReceiveClose() now always postTask the CLOSED transition
  and the CloseEvent dispatch, so onclose cannot run inside ws.close()/
  ws.terminate(). m_state moves to CLOSING synchronously first so send()/
  ping()/pong() called before the task runs take their CLOSING branch
  instead of reaching sendWebSocketData() with kind == None.
- wait_for_promise() bails when is_event_loop_alive() is false, matching
  wait_for_promise_with_termination(). Without it an unsettleable TLA
  promise spun auto_tick()/tick_without_idle() at 100% CPU.
- expect().resolves/.rejects, toThrow(async), and custom async matchers
  now throw when the promise stays pending instead of unreachable!().
- Tests moved from test/regression/issue/15665.test.ts to
  test/js/web/websocket/websocket-close-async-dispatch.test.ts and
  extended with a send()-after-close check.

Fixes #15665
