+++ src/jsc/bindings/webcore/WebSocket.cpp
@@ -1438,14 +1438,17 @@ void WebSocket::didReceiveClose(CleanStatus wasClean, unsigned short code, WTF::
         return;
     }
 
-    // Per WHATWG spec, "when the WebSocket connection is closed ... the
-    // user agent must queue a task" to set readyState to CLOSED and fire
-    // the close event. This path is reached synchronously from
-    // ws.terminate() (C++ → Zig cancel() → handleClose →
-    // dispatchAbruptClose → didFailWithErrorCode → here), so firing the
-    // event without a task would run onclose before terminate() returns.
-    // The Zig side nulls outgoing_websocket before calling back so no
-    // second callback can arrive while the task is pending.
+    // Per the WHATWG spec, "when the WebSocket connection is closed ... the
+    // user agent must queue a task" to set readyState to CLOSED and fire the
+    // close event. This path is reached synchronously from ws.terminate()
+    // (C++ terminate() -> Rust cancel() -> dispatch_abrupt_close ->
+    // didFailWithErrorCode -> here) and from socket-level failures while
+    // OPEN, so dispatching without a task would run onclose before the call
+    // returns. didFailWithErrorCode has already cleared
+    // m_connectedWebSocketKind, so move to CLOSING synchronously: send()/
+    // ping()/pong() called before the task runs then take their CLOSING
+    // branch instead of reaching sendWebSocketData() with kind == None.
+    m_state = CLOSING;
     if (auto* context = scriptExecutionContext()) {
         const bool dispatchError = wasConnecting && isConnectionError;
         this->incPendingActivityCount();
@@ -1518,15 +1521,18 @@ void WebSocket::didClose(unsigned unhandledBufferedAmount, unsigned short code, 
     // so we just call decPendingActivityCount() after dispatching the event
     ASSERT(m_pendingActivityCount > 0);
 
-    // Per WHATWG spec, "when the WebSocket connection is closed ... the
-    // user agent must queue a task" to set readyState to CLOSED and fire
-    // the close event. This is reached synchronously from ws.close()
-    // (C++ close() → Zig close() → sendCloseWithBody → dispatchClose →
-    // here), so dispatching without a task would run onclose before
-    // close() returns — observably wrong (issue #15665). The Zig side
-    // nulls outgoing_websocket before calling back, and
-    // m_connectedWebSocketKind is already None above, so no second
-    // callback can arrive while the task is pending.
+    // Per the WHATWG spec, "when the WebSocket connection is closed ... the
+    // user agent must queue a task" to set readyState to CLOSED and fire the
+    // close event. This is reached synchronously from ws.close() (C++
+    // close() -> Rust close() -> send_close_with_body -> dispatch_close ->
+    // here) and from a server-initiated Close, so dispatching without a task
+    // would run onclose before close() returns (issue #15665).
+    // m_connectedWebSocketKind is already None above and the Rust side has
+    // cleared outgoing_websocket, so no second callback can arrive; move to
+    // CLOSING synchronously so send()/ping()/pong() called before the task
+    // runs take their CLOSING branch instead of reaching sendWebSocketData()
+    // with kind == None.
+    m_state = CLOSING;
     if (auto* context = scriptExecutionContext()) {
         context->postTask([code, wasClean, reason, protectedThis = Ref { *this }](ScriptExecutionContext& context) {
             ASSERT(protectedThis->scriptExecutionContext());
@@ -1739,7 +1745,7 @@ void WebSocket::didFailWithErrorCode(Bun::WebSocketErrorCode code)
     }
     }
 
-    // didReceiveClose already set m_state = CLOSED. The connect() ref
+    // didReceiveClose has queued the CLOSED transition. The connect() ref
     // kept us alive across the switch (including across the native
     // onClose callback dropping its RefPtr); release it from a task so
     // the caller's stack frame unwinds first. ContextDestructionObserver
